requirements.txt
google-api-python-client>=2.129.0
google-auth>=2.34.0
gspread>=6.1.2
gspread-dataframe>=3.3.1
pandas>=2.2.2
numpy>=1.26.4
